Change vram_block type from int to uint8_t

+++ src/memory.h
@@ -11,7 +11,7 @@ extern uint8_t memory[MEM_SIZE];
 extern uint8_t* vram;
 extern uint8_t* oam;
 
-extern int vram_block;
+extern uint8_t vram_block;
 
 extern size_t vram_size;
 extern size_t oam_size;
